Remove special handling of .NOTPARALLEL and specify GNU make
The handling of the .NOTPARALLEL: pseudo-target was not compatible
with the original behavior in GNU make because it entirely disabled
parallel builds for recursive makes, not just at the toplevel.
This resulted in disabling parallel makes for some CMake projects,
such as webkitgtk.

To make it clear that you can just use .NOTPARALLEL, specify what
was already implied - that the make command used has to be
(sufficiently) compatible with GNU make. To be precise, we'd have
to have a list of extensions that the makefiles can count on, but
in practice, anybody relying on the build-api will just use GNU
make.